cosign dist working
